refactor(interceptor): import logger from `@zimic/utils`
