+++ packages/zimic-interceptor/src/cli/browser/init.ts
@@ -2,7 +2,7 @@ import fs from 'fs';
 import path from 'path';
 import color from 'picocolors';
 
-import { logWithPrefix } from '@/utils/console';
+import { logger } from '@/utils/logging';
 
 import { SERVICE_WORKER_FILE_NAME } from './shared/constants';
 
@@ -20,8 +20,7 @@ async function initializeBrowserServiceWorker({ publicDirectory }: BrowserServic
   const destinationPath = path.join(absolutePublicDirectory, SERVICE_WORKER_FILE_NAME);
   await fs.promises.copyFile(MOCK_SERVICE_WORKER_PATH, destinationPath);
 
-  logWithPrefix(`Service worker script saved to ${color.magenta(destinationPath)}!`);
-  logWithPrefix('You can now use browser interceptors!');
+  logger.info(`Service worker script saved to ${color.magenta(destinationPath)}.`);
 }
 
 export default initializeBrowserServiceWorker;
