+++ packages/zimic-interceptor/src/cli/server/token/remove.ts
@@ -1,7 +1,7 @@
 import color from 'picocolors';
 
 import { readInterceptorTokenFromFile, removeInterceptorToken } from '@/server/utils/auth';
-import { logWithPrefix } from '@/utils/console';
+import { logger } from '@/utils/logging';
 
 interface InterceptorServerCreateTokenOptions {
   tokenId: string;
@@ -12,11 +12,11 @@ export async function removeInterceptorServerToken({ tokenId, tokensDirectory }:
   const token = await readInterceptorTokenFromFile(tokenId, { tokensDirectory });
 
   if (!token) {
-    logWithPrefix(`${color.red(color.bold('✘'))} Token ${color.red(tokenId)} not found.`, { method: 'error' });
+    logger.error(`${color.red(color.bold('✘'))} Token ${color.red(tokenId)} not found.`);
     process.exit(1);
   }
 
   await removeInterceptorToken(tokenId, { tokensDirectory });
 
-  logWithPrefix(`${color.green(color.bold('✔'))} Token ${color.green(token.name ?? tokenId)} removed.`);
+  logger.info(`${color.green(color.bold('✔'))} Token ${color.green(token.name ?? tokenId)} removed.`);
 }
